Restructure consent integration tests
